fix #22 melhoria nos consumers
